% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_triplet_to_set_format}
\alias{convert_triplet_to_set_format}
\title{convert triplet to set format}
\usage{
convert_triplet_to_set_format(gene_miRNA_interactions)
}
\arguments{
\item{gene_miRNA_interactions}{a data frame with one gene-miRNA interaction per row}
}
\value{
a list of genes where each entry is a vector of miRNAs
}
\description{
convert triplet to set format
}
\examples{
convert_triplet_to_set_format(gene_mir_interactions_triplet)
}
