% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jami.R
\name{jami}
\alias{jami}
\title{Wrapper function for calling JAMI from R}
\usage{
jami(gene_miRNA_interactions, gene_expr, mir_expr, output_file = NULL)
}
\arguments{
\item{gene_miRNA_interactions}{either a path, a data frame or a list of interacions to consider}

\item{gene_expr}{either a path, a data frame or a matrix for gene expression data}

\item{mir_expr}{either a path, a data frame or a matrix for miRNA expression data}

\item{output_file}{optional output path for JAMI results}
}
\value{
the JAMI result as data frame
}
\description{
Wrapper function for calling JAMI from R
}
