% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{jami_create_settings_manager}
\alias{jami_create_settings_manager}
\title{Create a JAMI SettingsManager instance}
\usage{
jami_create_settings_manager()
}
\value{
a populated settings manager
}
\description{
Create a JAMI SettingsManager instance
}
