% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{jami_settings}
\alias{jami_settings}
\title{Set or get settings for my package}
\usage{
jami_settings(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set settings.}
}
\description{
Set or get settings for my package
}
\section{Supported settings}{

The following settings are supported
\itemize{
 \item{\code{a}}{(\code{numeric};1) The value of a }
 \item{\code{b}}{(\code{numeric};2) The value of b }
}
}

