% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{write_to_JAMI_set_format}
\alias{write_to_JAMI_set_format}
\title{write list of gene miRNA interactions to JAMI set format}
\usage{
write_to_JAMI_set_format(gene_miRNA_interactions, output_file)
}
\arguments{
\item{gene_miRNA_interactions}{a list of genes where each entry is a vector of miRNAs}

\item{output_file}{where the file should be written to}
}
\value{
nothing
}
\description{
write list of gene miRNA interactions to JAMI set format
}
\examples{
write_to_JAMI_set_format(gene_mir_interactions_set, tempfile())
}
